Extend rotation3D for working with numpy arrays
The rotation3D function was previously only able to work with lists by
means of two given 2D points. Now, the main arguments has been changed
to a single vertex, which may rotate about a specific point. Also, the
function is now able to work in 2D and 3D.

Additionally, normL2 has been changed to return `np.linalg.norm` instead
of 'np.linalg', which is module, not a function.