Rewrite the Vertex constructor
The Vertex class now handles a wider range of possibilities for
initializing instances. Before, the Vertex class required three
positional arguments, describing the coordinates in a three-dimensional
euclidean space. For example,

  >>> tetris.Vertex(1, 2, 3)

Even though this is kind of acceptable, it restrains initialization via
`numpy.array` or other entities. In fact, the Vertex class stores the
coordinates as `numpy.array`. So, why not allow the user to initialize
Vertex with `numpy.array`?

Thus, the Vertex class can now be initialized with:

  * three or less numerical elements for the coordinates. Undeclared
    axis values are initialized as zero;

  * list, tuple, or numpy.array with three or less numerical elements
    for each coordinate. Again, undeclared axis values are initialized
    as zero;

Summarizing, the Vertex class became more flexible and, perhaps,
contributes to a more readable end-user code.