Add an option to set a uniform cell size in Block
The Block class now provides a method for setting an uniform cell size
in the block: `set_cell_size`. See the following example:

  >>> b = tetris.Block()
  >>> b.set_vertices([v1, v2, v3, v4, v5, v6, v7, v8])
  >>> b.set_cell_size(0.05)  # in meters

This way, the cell size will be at most 0.05 meters in each direction.
We say 'at most' because we must provide an integer number of cells for
the block. In some cases, we may face a non-integer number that results
in the desired cell size. Thus, we round up to get an integer number,
which will result in a cell size slightly smaller than the defined.